WIP: Simplified interactor tree, also changed synchronizer API to take vanilla Flutter IME artifacts.
