+++ super_editor/lib/src/core/document_composer.dart
@@ -18,8 +18,8 @@ class DocumentComposer with ChangeNotifier {
   /// desired.
   DocumentComposer({
     DocumentSelection? initialSelection,
-    DocumentImeConfiguration? imeConfiguration,
-  })  : imeConfiguration = ValueNotifier(imeConfiguration ?? const DocumentImeConfiguration()),
+    SuperEditorImeConfiguration? imeConfiguration,
+  })  : imeConfiguration = ValueNotifier(imeConfiguration ?? const SuperEditorImeConfiguration()),
         _preferences = ComposerPreferences() {
     _streamController = StreamController<DocumentSelectionChange>.broadcast();
     selectionNotifier.addListener(_onSelectionChangedBySelectionNotifier);
@@ -115,7 +115,7 @@ class DocumentComposer with ChangeNotifier {
     _streamController.sink.add(_latestSelectionChange);
   }
 
-  final ValueNotifier<DocumentImeConfiguration> imeConfiguration;
+  final ValueNotifier<SuperEditorImeConfiguration> imeConfiguration;
 
   /// The current IME composing region, which signifies spans of text
   /// that the IME is thinking about changing.
